requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
black==19.10b0
boto3==1.14.5
botocore==1.17.5
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Django==3.0.7
django-decouple==2.1
django-phonenumber-field==4.0.0
django-storages==1.9.1
docutils==0.15.2
importlib-metadata==1.6.1
jmespath==0.10.0
Markdown==3.2.2
passlib==1.7.2
pathspec==0.8.0
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
regex==2020.5.14
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
zipp==3.1.0
